remove concat_rows
